Resolved login issue with keyboard on RPI3B+
Garbage or no characters were printed on console when using keyboard. Used to work well with ssh.
The solution for the issue was already in the initial post-image script and was handled in the initial defconfig file.
The defconfig file was modified to include the default post-image script and arguments to resolve this issue.
Refer: https://www.raspberrypi.org/documentation/configuration/uart.md